requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
construct==2.5.3
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.1
djangorestframework==3.14.0
h11==0.14.0
haystack==0.42
idna==3.4
Markdown==3.5.1
outcome==1.3.0.post0
packaging==23.1
pefile==2023.2.7
Pillow==10.1.0
PySocks==1.7.1
python-ptrace==0.9.8
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
